+++ __mocks__/pdfMock.js
@@ -1,0 +1,2 @@
+// Mocks pdfjs modules in jest environments
+module.exports = Object.create(null);