+++ src/types/pdf-viewer.ts
@@ -1,0 +1,4 @@
+export type PDFViewerClasses = {
+  pdfPage?: string;
+  pdfDocument?: string;
+};